Moved useGetVoucherInstructionsSteps hook implementation
